Move stores to store/{version} packages
