+++ pkg/controller/mastership/v2/watcher.go
@@ -7,7 +7,7 @@ package v2
 import (
 	"context"
 	configapi "github.com/onosproject/onos-api/go/onos/config/v2"
-	configuration "github.com/onosproject/onos-config/pkg/store/configuration/v2"
+	"github.com/onosproject/onos-config/pkg/store/v2/configuration"
 	"github.com/onosproject/onos-lib-go/pkg/errors"
 	"sync"
