+++ pkg/controller/transaction/v2/watcher.go
@@ -6,8 +6,8 @@ package v2
 
 import (
 	"context"
-	proposalstore "github.com/onosproject/onos-config/pkg/store/proposal/v2"
-	transactionstore "github.com/onosproject/onos-config/pkg/store/transaction/v2"
+	proposalstore "github.com/onosproject/onos-config/pkg/store/v2/proposal"
+	transactionstore "github.com/onosproject/onos-config/pkg/store/v2/transaction"
 	"sync"
 
 	configapi "github.com/onosproject/onos-api/go/onos/config/v2"
