+++ pkg/manager/manager.go
@@ -12,16 +12,16 @@ import (
 	mastershipcontroller "github.com/onosproject/onos-config/pkg/controller/mastership/v2"
 	proposalcontroller "github.com/onosproject/onos-config/pkg/controller/proposal/v2"
 	"github.com/onosproject/onos-config/pkg/controller/target"
-	proposal "github.com/onosproject/onos-config/pkg/store/proposal/v2"
+	"github.com/onosproject/onos-config/pkg/store/v2/proposal"
 
 	"github.com/onosproject/onos-config/pkg/controller/node"
 	"github.com/onosproject/onos-config/pkg/northbound/admin"
 	gnminb "github.com/onosproject/onos-config/pkg/northbound/gnmi/v2"
 	"github.com/onosproject/onos-config/pkg/pluginregistry"
 	sb "github.com/onosproject/onos-config/pkg/southbound/gnmi"
-	configuration "github.com/onosproject/onos-config/pkg/store/configuration/v2"
 	"github.com/onosproject/onos-config/pkg/store/topo"
-	transaction "github.com/onosproject/onos-config/pkg/store/transaction/v2"
+	"github.com/onosproject/onos-config/pkg/store/v2/configuration"
+	"github.com/onosproject/onos-config/pkg/store/v2/transaction"
 	"github.com/onosproject/onos-lib-go/pkg/certs"
 	"github.com/onosproject/onos-lib-go/pkg/northbound"
