+++ pkg/northbound/admin/admin.go
@@ -13,8 +13,8 @@ import (
 	"github.com/onosproject/onos-api/go/onos/config/admin"
 	configapi "github.com/onosproject/onos-api/go/onos/config/v2"
 	"github.com/onosproject/onos-config/pkg/pluginregistry"
-	configuration "github.com/onosproject/onos-config/pkg/store/configuration/v2"
-	transaction "github.com/onosproject/onos-config/pkg/store/transaction/v2"
+	"github.com/onosproject/onos-config/pkg/store/v2/configuration"
+	"github.com/onosproject/onos-config/pkg/store/v2/transaction"
 	"github.com/onosproject/onos-config/pkg/utils"
 	pathutils "github.com/onosproject/onos-config/pkg/utils/path"
 	"github.com/onosproject/onos-config/pkg/utils/tree"
