+++ pkg/northbound/admin/admin_test.go
@@ -16,8 +16,8 @@ import (
 	gnmitest "github.com/onosproject/onos-config/pkg/northbound/gnmi/test"
 	"github.com/onosproject/onos-config/pkg/pluginregistry"
 	plugintest "github.com/onosproject/onos-config/pkg/pluginregistry/test"
-	configuration "github.com/onosproject/onos-config/pkg/store/configuration/v2"
-	transaction "github.com/onosproject/onos-config/pkg/store/transaction/v2"
+	"github.com/onosproject/onos-config/pkg/store/v2/configuration"
+	"github.com/onosproject/onos-config/pkg/store/v2/transaction"
 	"github.com/onosproject/onos-config/pkg/utils/path"
 	"github.com/onosproject/onos-lib-go/pkg/errors"
 	"github.com/openconfig/gnmi/proto/gnmi"
