+++ pkg/northbound/gnmi/v2/get_test.go
@@ -11,7 +11,7 @@ import (
 	proposalcontroller "github.com/onosproject/onos-config/pkg/controller/proposal/v2"
 	transactioncontroller "github.com/onosproject/onos-config/pkg/controller/transaction/v2"
 	sb "github.com/onosproject/onos-config/pkg/southbound/gnmi"
-	proposal "github.com/onosproject/onos-config/pkg/store/proposal/v2"
+	proposal "github.com/onosproject/onos-config/pkg/store/v2/proposal"
 	"github.com/onosproject/onos-lib-go/pkg/controller"
 	"github.com/openconfig/gnmi/proto/gnmi_ext"
 	"os"
@@ -28,8 +28,8 @@ import (
 	topoapi "github.com/onosproject/onos-api/go/onos/topo"
 	gnmitest "github.com/onosproject/onos-config/pkg/northbound/gnmi/test"
 	plugintest "github.com/onosproject/onos-config/pkg/pluginregistry/test"
-	configuration "github.com/onosproject/onos-config/pkg/store/configuration/v2"
-	transaction "github.com/onosproject/onos-config/pkg/store/transaction/v2"
+	configuration "github.com/onosproject/onos-config/pkg/store/v2/configuration"
+	transaction "github.com/onosproject/onos-config/pkg/store/v2/transaction"
 	"github.com/onosproject/onos-config/pkg/utils"
 	"github.com/onosproject/onos-config/pkg/utils/path"
 	"github.com/openconfig/gnmi/proto/gnmi"
