+++ pkg/store/v2/configuration/configuration.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package v2
+package configuration
 
 import (
 	"context"
