+++ pkg/store/v2/proposal/store.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package v2
+package proposal
 
 import (
 	"context"
