+++ pkg/store/v2/transaction/store_test.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package v2
+package transaction
 
 import (
 	"context"
