+++ pkg/store/v3/configuration/store.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package v3
+package configuration
 
 import (
 	"context"
