+++ pkg/store/v3/configuration/store_test.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package v2
+package configuration
 
 import (
 	"context"
@@ -10,7 +10,7 @@ import (
 	"testing"
 	"time"
 
-	configapi "github.com/onosproject/onos-api/go/onos/config/v2"
+	configapi "github.com/onosproject/onos-api/go/onos/config/v3"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -24,15 +24,23 @@ func TestConfigurationStore(t *testing.T) {
 	store2, err := NewAtomixStore(cluster)
 	assert.NoError(t, err)
 
-	target1 := configapi.TargetID("target-1")
-	target2 := configapi.TargetID("target-2")
+	target1 := configapi.Target{
+		ID:      "target-1",
+		Type:    "foo",
+		Version: "1",
+	}
+	target2 := configapi.Target{
+		ID:      "target-2",
+		Type:    "bar",
+		Version: "1",
+	}
 
 	ch := make(chan configapi.ConfigurationEvent)
 	err = store2.Watch(context.Background(), ch)
 	assert.NoError(t, err)
 
-	target1ConfigValues := make(map[string]*configapi.PathValue)
-	target1ConfigValues["/foo"] = &configapi.PathValue{
+	target1ConfigValues := make(map[string]configapi.PathValue)
+	target1ConfigValues["/foo"] = configapi.PathValue{
 		Path: "/foo",
 		Value: configapi.TypedValue{
 			Bytes: []byte("Hello world!"),
@@ -41,13 +49,14 @@ func TestConfigurationStore(t *testing.T) {
 	}
 
 	target1Config := &configapi.Configuration{
-		ID:       configapi.ConfigurationID(target1),
-		TargetID: target1,
-		Values:   target1ConfigValues,
+		ID: configapi.ConfigurationID{
+			Target: target1,
+		},
 	}
+	target1Config.Committed.Values = target1ConfigValues
 
-	target2ConfigValues := make(map[string]*configapi.PathValue)
-	target2ConfigValues["/foo"] = &configapi.PathValue{
+	target2ConfigValues := make(map[string]configapi.PathValue)
+	target2ConfigValues["/foo"] = configapi.PathValue{
 		Path: "bar",
 		Value: configapi.TypedValue{
 			Bytes: []byte("Hello world again!"),
@@ -55,26 +64,27 @@ func TestConfigurationStore(t *testing.T) {
 		},
 	}
 	target2Config := &configapi.Configuration{
-		ID:       configapi.ConfigurationID(target2),
-		TargetID: target2,
-		Values:   target2ConfigValues,
+		ID: configapi.ConfigurationID{
+			Target: target2,
+		},
 	}
+	target2Config.Committed.Values = target2ConfigValues
 
 	err = store1.Create(context.TODO(), target1Config)
 	assert.NoError(t, err)
-	assert.Equal(t, configapi.ConfigurationID(target1), target1Config.ID)
+	assert.Equal(t, target1, target1Config.ID.Target)
 	assert.NotEqual(t, configapi.Revision(0), target1Config.Revision)
 
 	err = store2.Create(context.TODO(), target2Config)
 	assert.NoError(t, err)
-	assert.Equal(t, configapi.ConfigurationID(target2), target2Config.ID)
+	assert.Equal(t, target2, target2Config.ID.Target)
 	assert.NotEqual(t, configapi.Revision(0), target2Config.Revision)
 
 	// Get the configuration
-	target1Config, err = store2.Get(context.TODO(), configapi.ConfigurationID(target1))
+	target1Config, err = store2.Get(context.TODO(), configapi.ConfigurationID{Target: target1})
 	assert.NoError(t, err)
 	assert.NotNil(t, target1Config)
-	assert.Equal(t, configapi.ConfigurationID(target1), target1Config.ID)
+	assert.Equal(t, target1, target1Config.ID.Target)
 	assert.NotEqual(t, configapi.Revision(0), target1Config.Revision)
 
 	// Verify events were received for the configurations
@@ -103,7 +113,7 @@ func TestConfigurationStore(t *testing.T) {
 	assert.Equal(t, 2, len(configurationList))
 
 	// Read and then update the configuration
-	target2Config, err = store2.Get(context.TODO(), configapi.ConfigurationID(target2))
+	target2Config, err = store2.Get(context.TODO(), configapi.ConfigurationID{Target: target2})
 	assert.NoError(t, err)
 	assert.NotNil(t, target2Config)
 	target2Config.Status.State = configapi.ConfigurationStatus_SYNCHRONIZED
@@ -116,9 +126,9 @@ func TestConfigurationStore(t *testing.T) {
 	assert.Equal(t, target2Config.ID, event.Configuration.ID)
 
 	// Verify that concurrent updates fail
-	target1Config11, err := store1.Get(context.TODO(), configapi.ConfigurationID(target1))
+	target1Config11, err := store1.Get(context.TODO(), configapi.ConfigurationID{Target: target1})
 	assert.NoError(t, err)
-	target1Config12, err := store2.Get(context.TODO(), configapi.ConfigurationID(target1))
+	target1Config12, err := store2.Get(context.TODO(), configapi.ConfigurationID{Target: target1})
 	assert.NoError(t, err)
 
 	target1Config11.Status.State = configapi.ConfigurationStatus_SYNCHRONIZED
