+++ pkg/store/v3/transaction/store_test.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package v3
+package transaction
 
 import (
 	"context"
